+++ src/index.js
@@ -1,4 +1,4 @@
-import { onready } from './modules/main.js';
+import onready from './modules/app.js';
 import './index.css';
 
 window.addEventListener('DOMContentLoaded', onready);
